model-wrapper/requirements.txt
# Core ML/DL libraries
torch==2.1.2
pytorch-lightning==2.1.3
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# MLflow
mlflow==2.8.1

# Data processing
pyyaml==6.0.1
json5==0.9.14

# CLI and utilities
argparse
logging

# Optional: For advanced features
# feast==0.34.0  # Feature store
# tecton==0.7.0  # Alternative feature store
# langgraph==0.0.20  # LLM workflows
# openai==1.3.7  # LLM API
